refactor(cua): isolate strict MCP response decoding
Keep CUA frame bounds and fatal error ordering in a private decoder while
core retains request and process ownership. Replace the transport's this
alias with a lexical write-failure callback; no budget growth.

Proof: node scripts/run-vitest.mjs extensions/cua-computer/src/mcp-driver-client.test.ts
(16 passed); extension types and dead exports passed. pnpm build passed
in 6m24.7s.
Local extension lint remains blocked by the documented nested-checkout
declaration guard; hosted CI must verify it before landing.
